Refactor: Modularize salary components API
- I created a new route file `backend/routes/salaryComponents.js` for salary component API logic.
- I updated `backend/server.js` to use the new salary components router.
- I implemented authentication middleware for the salary components routes.
- I aligned `salaryComponents.test.js` with the integration test style, renaming it to `backend/__tests__/salaryComponents.integration.test.js` and updating its content.

This change separates concerns, improves code organization, and enhances testability for the salary components feature.
